# spec/openapi.yml
openapi: 3.0.3
info:
  title: micro-chirp API
  description: A highly available SNS application using Event Sourcing and CQRS
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /posts:
    post:
      tags:
        - posts
      operationId: postPosts
      summary: Create a new post
      description: Create a new post with text content (grapheme-based character limit validation)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                  description: User ID of the post author
                content:
                  type: string
                  description: Text content of the post
              required:
                - userId
                - content
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  postId:
                    type: string
                    format: uuid
                    description: ID of the created post
                  userId:
                    type: string
                    format: uuid
                    description: User ID of the post author
                  content:
                    type: string
                    description: Text content of the post
                  createdAt:
                    type: string
                    format: date-time
                    description: Post creation timestamp
                required:
                  - postId
                  - userId
                  - content
                  - createdAt
        '400':
          description: Bad request (e.g., content exceeds character limit)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
  /posts/{postId}:
    get:
      tags:
        - posts
      operationId: getPostsById
      summary: Get post details
      description: Retrieve detailed information about a specific post including like/repost/reply/view counts
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the post to retrieve
        - name: userId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Current user ID to check like/repost status
      responses:
        '200':
          description: Post details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  postId:
                    type: string
                    format: uuid
                    description: Post ID
                  userId:
                    type: string
                    format: uuid
                    description: User ID of the post author
                  content:
                    type: string
                    description: Text content of the post
                  createdAt:
                    type: string
                    format: date-time
                    description: Post creation timestamp
                  likeCount:
                    type: integer
                    description: Number of likes
                  repostCount:
                    type: integer
                    description: Number of reposts
                  replyCount:
                    type: integer
                    description: Number of replies
                  viewCount:
                    type: integer
                    description: Number of views
                  isLikedByCurrentUser:
                    type: boolean
                    description: Whether current user has liked this post
                  isRepostedByCurrentUser:
                    type: boolean
                    description: Whether current user has reposted this post
                required:
                  - postId
                  - userId
                  - content
                  - createdAt
                  - likeCount
                  - repostCount
                  - replyCount
                  - viewCount
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message

    delete:
      tags:
        - posts
      operationId: deletePostsById
      summary: Delete a post
      description: Delete an existing post (soft delete via event)
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the post to delete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                  description: User ID requesting the deletion
              required:
                - userId
      responses:
        '204':
          description: Post deleted successfully
        '400':
          description: Bad request (e.g., user not found)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
        '403':
          description: Forbidden (user is not the post author)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
...
  /posts/{postId}/reposts:
    post:
      tags:
        - reposts
      operationId: postReposts
      summary: Repost a post
      description: Add a repost to a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the post to repost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                  description: User ID performing the repost
              required:
                - userId
      responses:
        '201':
          description: Post reposted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  postId:
                    type: string
                    format: uuid
                    description: Post ID
                  userId:
                    type: string
                    format: uuid
                    description: User ID who reposted the post
                  repostedAt:
                    type: string
                    format: date-time
                    description: Repost timestamp
                required:
                  - postId
                  - userId
                  - repostedAt
        '400':
          description: Bad request (e.g., user not found)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message

    delete:
      tags:
        - reposts
      operationId: deleteReposts
      summary: Unrepost a post
      description: Remove a repost from a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the post to unrepost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                  description: User ID performing the unrepost
              required:
                - userId
      responses:
        '204':
          description: Post unreposted successfully
        '400':
          description: Bad request (e.g., user not found)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
...